# backend/api/swagger.yaml
openapi: 3.0.1
info:
  description: A simple application to handle tasks.
  license:
    name: Eclipse Public License 2.0
    url: https://www.eclipse.org/org/documents/epl-2.0/EPL-2.0.html
  title: A TODO-Task list application
  version: 1.0.0
servers:
- url: http://localhost:8080/todoapp
tags:
- description: Task management
  name: task
paths:
  /user:
    get:
      operationId: getUser
      responses:
        "200":
          description: sucess
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
  /task:
    get:
      operationId: tasksGetAll
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Task'
                type: array
          description: List of all tasks
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Generic error response
      summary: Get the list of all tasks
      tags:
      - task
    post:
      operationId: tasksCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: The created task
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Generic error response
      summary: Create a new task
      tags:
      - task
  /task/{taskId}:
    delete:
      operationId: tasksDelete
      parameters:
      - description: The id of the task
        example: e1cb23d0-6cbe-4a29-b586-bfa424bc93fd
        explode: false
        in: path
        name: taskId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Task deleted response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Generic error response
      summary: Delete an existing task
      tags:
      - task
    get:
      operationId: tasksRead
      parameters:
      - description: The id of the task
        example: e1cb23d0-6cbe-4a29-b586-bfa424bc93fd
        explode: false
        in: path
        name: taskId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Ok
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Generic error response
      summary: Get a single task based on its id
      tags:
      - task
    put:
      operationId: tasksUpdate
      parameters:
      - description: The id of the task
        example: e1cb23d0-6cbe-4a29-b586-bfa424bc93fd
        explode: false
        in: path
        name: taskId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: The updated task
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Generic error response
      summary: Update an existing task
      tags:
      - task
components:
  schemas:
    User:
      description: User
      properties:
        id:
          type: string
          readOnly: true
        name:
          type: string
    Task:
      description: Object representing a Task
      example:
        description: My important task
        completed: true
        id: e1cb23d0-6cbe-4a29-b586-bfa424bc93fd
      properties:
        completed:
          description: indicates if a taks is completed or not
          type: boolean
        description:
          description: description of the task
          example: My important task
          type: string
        id:
          description: id of the taks
          example: e1cb23d0-6cbe-4a29-b586-bfa424bc93fd
          readOnly: true
          type: string
      required:
      - description
      type: object
    Error:
      description: Object representing an error
      example:
        code: 1000
        message: Could not perform the task
      properties:
        code:
          description: Error code that identify of the error
          example: 1000
          format: int32
          type: integer
        message:
          description: Short description of the error
          example: Could not perform the task
          type: string
      type: object